performance:
```
cpu: Intel(R) Xeon(R) Platinum 8260 CPU @ 2.40GHz
                                         │ /tmp/tmpw47pipaa.source.txt │     /tmp/tmpx034z3w7.target.txt      │
                                         │           sec/op            │    sec/op     vs base                │
UnmarshalConcrete/CanadaGeometry_Sonic-8                  1.027m ± 12%   1.071m ±  5%        ~ (p=0.353 n=10)
UnmarshalConcrete/CitmCatalog_Sonic-8                     6.561m ±  6%   6.913m ±  6%        ~ (p=0.089 n=10)
UnmarshalConcrete/GolangSource_Sonic-8                    16.46m ±  5%   16.27m ±  4%        ~ (p=0.684 n=10)
UnmarshalConcrete/StringEscaped_Sonic-8                  106.76µ ±  6%   92.77µ ± 11%  -13.10% (p=0.000 n=10)
UnmarshalConcrete/StringUnicode_Sonic-8                   10.90µ ±  5%   10.85µ ±  6%        ~ (p=0.218 n=10)
UnmarshalConcrete/SyntheaFhir_Sonic-8                     12.86m ±  6%   12.72m ±  5%        ~ (p=0.912 n=10)
UnmarshalConcrete/TwitterStatus_Sonic-8                   2.608m ±  9%   2.556m ±  7%        ~ (p=0.353 n=10)
MarshalConcrete/CanadaGeometry_Sonic-8                    923.8µ ±  4%   909.9µ ±  5%        ~ (p=0.631 n=10)
MarshalConcrete/CitmCatalog_Sonic-8                       919.9µ ±  9%   736.1µ ±  1%  -19.98% (p=0.000 n=10)
MarshalConcrete/GolangSource_Sonic-8                      3.458m ±  4%   3.052m ±  5%  -11.74% (p=0.000 n=10)
MarshalConcrete/StringEscaped_Sonic-8                     7.292µ ±  8%   7.537µ ±  6%        ~ (p=0.289 n=10)
MarshalConcrete/StringUnicode_Sonic-8                     7.049µ ±  4%   7.367µ ±  3%   +4.50% (p=0.019 n=10)
MarshalConcrete/SyntheaFhir_Sonic-8                       12.68m ±  5%   12.56m ±  5%        ~ (p=0.280 n=10)
MarshalConcrete/TwitterStatus_Sonic-8                     419.7µ ± 13%   389.8µ ±  6%   -7.13% (p=0.003 n=10)
geomean                                                   667.4µ         645.7µ         -3.25%
```